refactor common verification logic for sauce and phantom
